replace current with radix implementations
