+++ packages/web-ui/src/FieldsetLegend/FieldsetLegend.tsx
@@ -20,17 +20,27 @@ export interface FieldsetLegendProps extends SxProps, HTMLAttributes<HTMLLegendE
  * grouped from inputs.
  */
 export const FieldsetLegend = forwardRef<HTMLLegendElement, FieldsetLegendProps>(
-  ({ disabled, ...props }, ref) => {
+  ({ disabled, sx, ...props }, ref) => {
+    const defaultColor = colors.grey1000;
+    const disabledColor = colors.grey400;
     return (
       <Box
         ref={ref}
         component="legend"
         padding={0} // reset
-        color={disabled ? colors.grey400 : colors.grey1000}
+        color={disabled ? disabledColor : defaultColor}
         fontFamily={fonts.secondary}
         fontWeight={fontWeights.secondary.semibold}
         fontSize={pxToRem(16)}
         lineHeight={pxToRem(16)}
+        sx={{
+          cursor: disabled ? 'auto' : 'pointer',
+          color: defaultColor,
+          '[data-disabled] &': {
+            color: disabledColor,
+          },
+          ...sx,
+        }}
         {...props}
       />
     );
