+++ packages/web-ui/src/RadioGroup/Radio.stories.tsx
@@ -17,50 +17,42 @@ export const RadioStory: Story = {
   name: 'Radio',
   render: args => {
     return (
-      <Box bgcolor="white" padding={4}>
-        <Stack spacing={4}>
-          <RadioGroup value="2" label="Unchecked Radio">
-            <Radio {...args} />
-          </RadioGroup>
+      <Stack spacing={4}>
+        <RadioGroup value="2" label="Unchecked Radio">
+          <Radio {...args} />
+        </RadioGroup>
 
-          <RadioGroup defaultValue={args.value} label="Checked Radio">
-            <Radio {...args} />
-          </RadioGroup>
-        </Stack>
-      </Box>
+        <RadioGroup defaultValue={args.value} label="Checked Radio">
+          <Radio {...args} />
+        </RadioGroup>
+      </Stack>
     );
   },
   argTypes: {
     value: { control: { type: 'text' } },
     helperText: { control: { type: 'text' } },
-    children: { control: { type: 'text' } },
+    label: { control: { type: 'text' } },
     disabled: { control: { type: 'boolean' } },
   },
   args: {
     value: '1',
     disabled: false,
-    children: 'Label',
+    label: 'Label',
     helperText: 'Helper text',
   },
 };
 
-export const WithCustomRadio: Story = {
+export const WithCustomLabel: Story = {
   name: 'Custom Radio label',
   render: () => {
     const CustomLabel = (props: { children: TextProps['children'] }) => (
       <Text {...props} variant="subtitle" component="span" bold />
     );
     return (
-      <Box bgcolor="white" padding={4}>
-        <RadioGroup defaultValue="1" label="Radio group label">
-          <Radio value="1">
-            <CustomLabel>One</CustomLabel>
-          </Radio>
-          <Radio value="2">
-            <CustomLabel>Two</CustomLabel>
-          </Radio>
-        </RadioGroup>
-      </Box>
+      <RadioGroup defaultValue="1" label="Radio group label">
+        <Radio value="1" label={<CustomLabel>One</CustomLabel>} />
+        <Radio value="2" label={<CustomLabel>Two</CustomLabel>} />
+      </RadioGroup>
     );
   },
 };
