+++ packages/web-ui/src/RadioGroup/RadioGroup.docs.mdx
@@ -1,7 +1,6 @@
 import { Meta, Description, Canvas, ArgTypes } from '@storybook/blocks';
 import * as RadioGroupStories from './RadioGroup.stories';
 import * as RadioStories from './Radio.stories';
-import * as RadioTileStories from './RadioTile.stories';
 import { Stack } from '../Stack';
 import { Box } from '../Box';
 import { Text } from '../Text';
@@ -11,53 +10,25 @@ import { Divider } from '@mui/material';
 
 # RadioGroup
 
+<Description of={RadioGroupStories} />
+
 - [Accessibility](#accessibility)
   - [Known Issues](#known-issues)
   - [Keyboard interactions](#keyboard-interactions)
 - [Helper text](#helper-text)
 - [Error message](#error-message)
 - [Width](#width)
-- [Columns](#columns)
 - [Controlled](#controlled)
 - [RadioGroup props](#radiogroup-props)
-- [Radio](#radio)
-  - [Radio accessibility](#radio-accessibility)
-  - [Radio props](#radio-props)
-- [RadioTile](#radiotile)
-  - [RadioTile accessibility](#radiotile-accessibility)
-  - [RadioTile props](#radiotile-props)
+- [Radio & RadioTile](#radio-&-radiotile)
+  - [Accessibility](#accessibility)
+  - [Props](#props)
 
-<Description of={RadioGroupStories} />
-
-<Canvas
-  of={RadioGroupStories.Workshop}
-  sourceState="shown"
-  source={{
-    code: `
-<RadioGroup defaultValue={1} label="Label" helperText="RadioGroup Helper text">
-  <Radio value="1" helperText="One helper text">
-    One
-  </Radio>
-  <Radio value="2" helperText="Two helper text">
-    Two
-  </Radio>
-  <Radio value="3" helperText="Three helper text">
-    Three
-  </Radio>
-</RadioGroup>
-`,
-  }}
-/>
+<Canvas of={RadioGroupStories.Workshop} sourceState="shown" />
 
 ## Accessibility
 
 Follows the [WAI-ARIA Radio Group Pattern](https://www.w3.org/WAI/ARIA/apg/patterns/radio/) for radio groups not contained in a toolbar.
-
-> **NOTE**
-> The initial focus behaviour when using `Shift+Tab` to move focus into the
-> radio group, if none of the radio buttons are selected, reflects the native
-> browser behaviour, rather than that described in the WAI-ARIA pattern, as it
-> focuses the last radio button in the group, rather than the first.
 
 The necessary aria props, such as `aria-labelledby`, `aria-describedby` and
 `aria-errormessage` are handled internally, however you can pass these and
@@ -149,14 +120,6 @@ the `RadioGroup`.
 This is a responsive property, so you are able to set different widths for
 different breakpoints.
 
-## Columns
-
-You can use the `columns` prop to display the RadioGroup's child elements in
-columns rather than a flat column or row, by setting the number of columns you
-need.
-
-<Canvas of={RadioGroupStories.Columns} />
-
 ## Controlled
 
 A controlled value can be provided using the value prop, which accepts a value
@@ -183,8 +146,6 @@ const [selected, setSelected] = useState('')
 
 ## RadioGroup props
 
-<ArgTypes of={RadioGroupStories} />
-
 ## Radio & RadioTile
 
 The `Radio` & `RadioTile` components should be used within a `RadioGroup` component.
@@ -197,4 +158,4 @@ accessibility.
 
 ### Props
 
-<ArgTypes of={RadioTileStories} />
+<ArgTypes of={RadioStories} />
