+++ packages/web-ui/src/RadioGroup/RadioGroup.stories.tsx
@@ -1,6 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { Radio } from './Radio';
-import { RadioTile } from './RadioTile';
 import { RadioGroup } from './RadioGroup';
 import { Stack } from '../Stack';
 import { useState } from 'react';
@@ -8,57 +7,45 @@ import { useState } from 'react';
 const meta: Meta<typeof RadioGroup> = {
   title: 'Web UI / Components / RadioGroup',
   component: RadioGroup,
+  argTypes: {
+    direction: {
+      options: ['column', 'row'],
+      control: { type: 'radio' },
+    },
+    error: { control: { type: 'boolean' } },
+    defaultValue: { control: { type: 'text' } },
+    helperText: { control: { type: 'text' } },
+    helperTextPosition: { options: ['top', 'bottom'], control: { type: 'radio' } },
+    label: { control: { type: 'text' } },
+    errorMessage: { control: { type: 'text' } },
+    disabled: { control: { type: 'boolean' } },
+    contentWidth: { control: { type: 'text' } },
+  },
+  args: {
+    defaultValue: '1',
+    label: 'Radio group label',
+    helperText: 'Helper text',
+    contentWidth: undefined,
+    disabled: false,
+    error: false,
+    errorMessage: 'There is an error',
+  },
 };
 
 export default meta;
 type Story = StoryObj<typeof RadioGroup>;
 
-const argTypes = {
-  direction: {
-    options: ['column', 'row'],
-    control: { type: 'radio' },
-  },
-  error: { control: { type: 'boolean' } },
-  defaultValue: { control: { type: 'text' } },
-  helperText: { control: { type: 'text' } },
-  helperTextPosition: { options: ['top', 'bottom'], control: { type: 'radio' } },
-  label: { control: { type: 'text' } },
-  errorMessage: { control: { type: 'text' } },
-  disabled: { control: { type: 'boolean' } },
-  isReadOnly: { control: { type: 'boolean' } },
-  contentWidth: { control: { type: 'text' } },
-  columns: { control: { type: 'number' } },
-};
-const sharedArgs = {
-  label: 'Radio group label',
-  helperText: 'Helper text',
-  defaultValue: '1',
-  contentWidth: undefined,
-  disabled: false,
-  error: false,
-  errorMessage: 'There is an error',
-};
-
 export const Workshop: Story = {
   name: 'RadioGroup',
   render: args => {
     return (
-      <Stack spacing={8}>
-        <RadioGroup {...args} helperText="RadioGroup with Radio">
-          <Radio value="1">One</Radio>
-          <Radio value="2">Two</Radio>
-          <Radio value="3">Three</Radio>
-        </RadioGroup>
-        <RadioGroup {...args} helperText="RadioGroup with RadioTile">
-          <RadioTile value="1">One</RadioTile>
-          <RadioTile value="2">Two</RadioTile>
-          <RadioTile value="3">Three</RadioTile>
-        </RadioGroup>
-      </Stack>
+      <RadioGroup {...args} helperText="RadioGroup with Radio">
+        <Radio value="1" label="One" />
+        <Radio value="2" label="Two" />
+        <Radio value="3" label="Three" />
+      </RadioGroup>
     );
   },
-  argTypes,
-  args: sharedArgs,
 };
 
 export const WithRadioHelperText: Story = {
@@ -67,61 +54,33 @@ export const WithRadioHelperText: Story = {
     return (
       <Stack spacing={8}>
         <RadioGroup {...args}>
-          <Radio value="1" helperText="One helper text">
-            One
-          </Radio>
-          <Radio value="2" helperText="Two helper text">
-            Two
-          </Radio>
-          <Radio value="3" helperText="Three helper text">
-            Three
-          </Radio>
-        </RadioGroup>
-        <RadioGroup {...args}>
-          <RadioTile value="1" helperText="One helper text">
-            One
-          </RadioTile>
-          <RadioTile value="2" helperText="Two helper text">
-            Two
-          </RadioTile>
-          <RadioTile value="3" helperText="Three helper text">
-            Three
-          </RadioTile>
+          <Radio value="1" label="One" helperText="One helper text" />
+          <Radio value="2" label="Two" helperText="Two helper text" />
+          <Radio value="3" label="Three" helperText="Three helper text" />
         </RadioGroup>
       </Stack>
     );
-  },
-  argTypes,
-  args: {
-    defaultValue: '1',
-    disabled: false,
-    error: false,
-    label: 'Radio group',
   },
 };
 
 export const Controlled: Story = {
   render: args => {
     const [selected, setSelected] = useState('');
-
     return (
       <RadioGroup
         {...args}
         value={selected}
-        onChange={setSelected}
+        onValueChange={setSelected}
         helperText={`The selected value is: ${selected}`}
       >
-        <Radio value="1">One</Radio>
-        <Radio value="2">Two</Radio>
-        <Radio value="3">Three</Radio>
+        <Radio value="1" label="One" />
+        <Radio value="2" label="Two" />
+        <Radio value="3" label="Three" />
       </RadioGroup>
     );
   },
-  argTypes,
   args: {
     errorMessage: 'There is an error',
-    disabled: false,
-    error: false,
     label: 'Controlled Radio Group',
   },
 };
@@ -131,30 +90,14 @@ export const ShowingError: Story = {
   render: args => {
     return (
       <RadioGroup {...args}>
-        <Radio value="1">One</Radio>
-        <Radio value="2">Two</Radio>
-        <Radio value="3">Three</Radio>
+        <Radio value="1" label="One" />
+        <Radio value="2" label="Two" />
+        <Radio value="3" label="Three" />
       </RadioGroup>
     );
   },
-  argTypes,
-  args: { ...sharedArgs, error: true, errorMessage: 'Radio group error message' },
-};
-
-export const Columns: Story = {
-  name: 'RadioGroup columns',
-  render: args => {
-    return (
-      <RadioGroup {...args} label="RadioGroup with columns">
-        <RadioTile value="1">One</RadioTile>
-        <RadioTile value="2">Two</RadioTile>
-        <RadioTile value="3">Three</RadioTile>
-        <RadioTile value="4">Four</RadioTile>
-        <RadioTile value="5">Five</RadioTile>
-        <RadioTile value="6">Six</RadioTile>
-      </RadioGroup>
-    );
+  args: {
+    error: true,
+    errorMessage: 'Radio group error message',
   },
-  argTypes,
-  args: { columns: 2, ...sharedArgs },
 };
