+++ packages/web-ui/src/RadioGroup/index.ts
@@ -1,5 +1,7 @@
 export { RadioGroup } from './RadioGroup';
 export type { RadioGroupProps } from './RadioGroup';
+export { RadioGridGroup } from './RadioGridGroup';
+export type { RadioGridGroupProps } from './RadioGridGroup';
 export { Radio } from './Radio';
 export type { RadioProps } from './Radio';
 export { RadioTile } from './RadioTile';
